% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_meas_error.R
\name{apply_meas_error}
\alias{apply_meas_error}
\title{Measurement error}
\usage{
apply_meas_error(tru.val = NULL, max.val = Inf, coefVar = 0.05)
}
\arguments{
\item{tru.val}{True value vector}

\item{max.val}{Maximum value (upper bound of rtruncnorm function)}

\item{coefVar}{coefficient of variation of error spread around mean}
}
\description{
Apply measurement error
}
