% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioavailability-package.R
\docType{package}
\name{bioavailability-package}
\alias{bioavailability}
\alias{bioavailability-package}
\title{bioavailability: Estimating error guideline exceedance and compliance with bioavailabilty}
\description{
Estimates error guideline exceedance and compliance, based on user inputs, accounting for bioavailability
}
\author{
\strong{Maintainer}: Clay Nelson \email{claymnelson@gmail.com@gmail.com}

Authors:
\itemize{
  \item Kevin Li \email{kli.kevin@gmail.com}
}

}
\keyword{internal}
