% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx.R
\name{fx}
\alias{fx}
\title{Direct model, from mean measured IVBA to estimated RBA}
\usage{
fx(x, contaminant)
}
\arguments{
\item{x}{measured IVBA}

\item{contaminant}{string, "As" or "Pb"}
}
\value{
Estimated RBA value based on measured IVBA \code{x}
}
\description{
Direct model, from mean measured IVBA to estimated RBA
}
\examples{
fx(75, "Pb")

}
