% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy.R
\name{fy}
\alias{fy}
\title{Inverse model to calculate observed IVBA from true RBA without model error}
\usage{
fy(y, contaminant)
}
\arguments{
\item{y}{"true" RBA}

\item{contaminant}{string, "As" or "Pb"}
}
\value{
Estimated IVBA value based on RBA defined by \code{y}
}
\description{
Calculates IVBA based on provided RBA without model error.
}
\examples{
fy_error(65, "As")

}
