% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_error.R
\name{fy_error}
\alias{fy_error}
\title{Inverse model to calculate observed IVBA from true RBA with model error}
\usage{
fy_error(y, contaminant, sepred = NULL)
}
\arguments{
\item{y}{"true" RBA}

\item{contaminant}{string, "As" or "Pb"}

\item{sepred}{(Optional) a custom \code{sepred} value. Default is to use the
sepred value in \code{contam_params} object for \code{contaminant} "As" or "Pb"}
}
\value{
Estimated IVBA value based on RBA defined by \code{y} with error
}
\description{
Inverse model to calculate observed IVBA from true RBA with model error
}
\examples{
fy_error(65, "As")

}
