% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitewide_results.R
\name{hist.meas_rba_site}
\alias{hist.meas_rba_site}
\title{Plotting functions for sitewide app}
\usage{
\method{hist}{meas_rba_site}(site_error, tru.mean)
}
\arguments{
\item{site_error}{site error table}

\item{tru.mean}{true mean value set by user}
}
\value{
a plot
}
\description{
Plotting functions for sitewide app
}
