% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Find interpolated value}
\usage{
interp(x, y, middle.point.y)
}
\arguments{
\item{middle.point.y}{}
}
\value{
Interpolated value between the first and second values of \code{x}
}
\description{
Function to find interpolated value given two points (err_pb is the y)
}
\examples{
a=c(1,4)
b=c(4,6)
interp(a,b,5)
}
