% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDist_tot.R
\name{simDist_tot}
\alias{simDist_tot}
\title{Function to generate "true" total conc.}
\usage{
simDist_tot(n.totmeas, tru_mu_tot, coeV_tot, dist_tot)
}
\arguments{
\item{n.totmeas}{Number of total measurements}

\item{tru_mu_tot}{"True" mean total concentration}

\item{coeV_tot}{coefficient of variation of total concentration}

\item{dist_tot}{distribution of total concentration measurements; options are
"normal" or "lognorm"}
}
\value{
A vector of simulated total concentration measurements
}
\description{
Function to generate "true" total conc.
}
