% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSite.R
\name{simSite}
\alias{simSite}
\title{Simulate site function}
\usage{
simSite(
  DU.n,
  sample.n,
  mn_rba_site,
  coeV_rba_site,
  simDist_rba_site,
  ivba.incr,
  error_ivb = TRUE,
  ivba_model = FALSE,
  post_mean = FALSE,
  error_ivb_cv,
  iter = 5000
)
}
\arguments{
\item{DU.n}{number of DUs}

\item{sample.n}{number of samples measuring IVBA}

\item{mn_rba_site}{mean "true" RBA for the whole site}

\item{coeV_rba_site}{coefficient of varation of RBA for the whole site}

\item{simDist_rba_site}{distribution of DU RBA value across the site}

\item{ivba.incr}{number of increments per IVBA sample}

\item{error_ivb}{(TRUE/FALSE) apply IVBA measurement error? Default is FALSE.}

\item{ivba_model}{(TRUE/FALSE) apply IVBA model error? Default is FALSE.}

\item{post_mean}{(TRUE/FALSE) calculate IVBA model error after
summarizing across samples. Default is FALSE.}

\item{error_ivb_cv}{IVBA measurement error coefficient of variance}

\item{iter}{Number of simulation iterations}

\item{useMeanIVBA}{(TRUE/FALSE) use the mean IVBA in calculating error results?}
}
\value{
simulation results
}
\description{
Simulate site function
}
