% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1.R
\name{step1}
\alias{step1}
\title{Step 1 function}
\usage{
step1(
  AsPb = NULL,
  tot.n = NULL,
  ivba.n = NULL,
  sampmax = NULL,
  incr.vec = NULL,
  ...
)
}
\arguments{
\item{AsPb}{define contaminant}

\item{tot.n}{number of total conc. samples, passed to simDU}

\item{ivba.n}{number of ivba samples, passed to simDU}

\item{sampmax}{maximum number of samples to simulate, default 10*tot.n}

\item{incr.vec}{vector of increments to test}

\item{...}{simDU parameters}
}
\description{
Determine sequence of sample sizes to simulate
}
\examples{
test.step1 <- step1(AsPb = "Pb", tot.n = 5, ivba.n = 3, sampmax = 50,
frcAct = -0.25, coeV.tot = 0.5, coeV.rba = 0.3, error_tot = TRUE,
error_ivb = TRUE, error_ivb_cv = 0.05, ivba_model = TRUE, incr.vec = c(1,5))

}
