% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_plot.R
\name{step1_plot}
\alias{step1_plot}
\title{Step 1 plot}
\usage{
step1_plot(
  step1.output,
  error_threshold,
  sm.text = 14,
  lg.text = 16,
  plot.text = 4,
  guide.text = 14
)
}
\arguments{
\item{step1.output}{Output of \code{step1} function}

\item{error_threshold}{Desired error threshold line in plot}
}
\value{
A \code{ggplot2} plot object
}
\description{
A GUI function.
}
