% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1a_plot.R
\name{step1a_plot}
\alias{step1a_plot}
\title{Step1a plot}
\usage{
step1a_plot(step1a.output, sm.text = 14, lg.text = 16, hatch.spacing = 0.01)
}
\arguments{
\item{step1a.output}{Output of step1a.}
}
\value{
A \code{ggplot2} plot object
}
\description{
A GUI function.
}
