% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_plot.R
\name{step2_plot}
\alias{step2_plot}
\title{Step 2 plot}
\usage{
step2_plot(step2.output, error_threshold, sm.text = 14, lg.text = 16)
}
\arguments{
\item{step2.output}{Output of \code{step2} function}

\item{error_threshold}{Desired error threshold}
}
\value{
a \code{ggplot2} plot object
}
\description{
A GUI function
}
