% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_text.R
\name{step2_text_t1}
\alias{step2_text_t1}
\alias{step2_text_t2}
\title{Step 2 text functions}
\usage{
step2_text_t1(step2_t1, error_threshold)

step2_text_t2(step2_t2, error_threshold)
}
\arguments{
\item{step2_t1}{Type 1 error results}

\item{error_threshold}{Desired error threshold}

\item{step2_t2}{Type 2 error results}
}
\value{
A text string.
}
\description{
Step 2 text functions
}
