% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4.R
\name{step4}
\alias{step4}
\title{Step 4}
\usage{
step4(
  meas.tot = NULL,
  meas.ivba = NULL,
  AsPb = NULL,
  tot.incr = NULL,
  ivba.incr = NULL,
  useMeanTot = T,
  actLvl = 400,
  actLvlRBA = 60,
  ...
)
}
\arguments{
\item{meas.tot}{actual total concentration measurements}

\item{meas.ivba}{actual ivba measurements}

\item{AsPb}{"As" or "Pb"}

\item{tot.incr}{Sampling protocol: number of increments for total conc. samples}

\item{ivba.incr}{Sampling protocol: number of increments for ivb samples}

\item{useMeanTot}{Use mean of upper 95\% interval in estimation of total contaminant? T = use the mean; F = use 95\% interval}

\item{actLvl}{What is the site-specific soil contaminant action level? Action level: limit of biovalability above which site has to be remedied (mg/kg)}

\item{actLvlRBA}{\% RBA assumed in the action level}

\item{...}{\code{simDU} inputs}
}
\description{
Calculate precision and accuracy
}
\examples{
# teststep4 <- step4(testcv(rnorm(n = 100, mean = 400, sd = 39),10),
      meas.ivba = testcv(rnorm(n = 100, mean = 78, sd = 15),10),
      tot.incr = 10, ivba.incr = 10, "Pb")
}
