% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step4_plot.R
\name{step4_plot}
\alias{step4_plot}
\title{Step 4 plot}
\usage{
step4_plot(step4.output, sm.text = 14, lg.text = 16, strp.text = 20)
}
\arguments{
\item{step4.output}{output of \code{step4} function}
}
\value{
A list containing the output plot, an accuracy text string, and a
precision text string
}
\description{
Step 4 plot
}
