% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_samples.R
\name{take_samples}
\alias{take_samples}
\title{Take samples}
\usage{
take_samples(n.samp, n.incr, n.sim, distfun, ...)
}
\arguments{
\item{n.samp}{Number of samples}

\item{n.incr}{Number of increments}

\item{n.sim}{Number of simulations}

\item{distfun}{Distribution function}

\item{...}{Additional arguments to the distribution function}
}
\value{
Dataframe with simulation values and attributes.
}
\description{
Take samples
}
\examples{
take_samples(5, 2, 20, "simDist_rba", 50, 0.3, "lognorm")
}
