% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcv.R
\name{testcv}
\alias{testcv}
\title{Test the coev back calculation}
\usage{
testcv(x, incr)
}
\arguments{
\item{x}{vector of increment samples, values represent individual increments}

\item{incr}{numeric value of increment size, increment grouping size of
samples}
}
\value{
vector of "composited" samples, number of samples = x/incr
}
\description{
Function returns incremented x by taking means in groups of 'incr'. Used to
demonstrate that as effective number of "cores" increases (i.e., base unit of
sampling whether composited or not), e estimated coeV gets closer to the
original inputted coeV (sd=20, mn = 300)
}
