% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper95.R
\name{upper95}
\alias{upper95}
\title{Upper 95\% conf. int. of the mean function}
\usage{
upper95(x, lvl)
}
\arguments{
\item{x}{observation values}

\item{lvl}{desired quantile of \code{x} to calculate}
}
\description{
This is a convenience function for calculating the 95\% quantile of a vector
of observations \code{x}, intended for calculating the upper 95\% confidence value
}
\examples{
upper95(x=c(29,39,48,34,58), lvl=0.95)
}
