% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDist_rba.R
\name{simDist_rba}
\alias{simDist_rba}
\title{Function to generate "true" rba}
\usage{
simDist_rba(n.rbameas, tru_mu_rba, coeV_rba, dist_rba)
}
\arguments{
\item{n.rbameas}{Number of RBA measurements}

\item{tru_mu_rba}{Mean "true" RBA value}

\item{coeV_rba}{Coefficient of variation of RBA distribution}

\item{dist_rba}{Distribution of RBA measurements; can be "normal", "lognorm",
or "uniform"}
}
\value{
A vector of simulated RBA measurements
}
\description{
Function to generate "true" rba
}
